@@ -13,9 +13,7 @@ namespace DoePaAdminDataModel.Kostenrechnung
         public Kostenstelle ZugehoerigeKostenstelle { get; set; }
 
         public decimal Stueckzahl { get; set; }
-
-        public int? ZugehoerigeAbrechnungseinheitID { get; set; }
-
+        
         public Abrechnungseinheit ZugehoerigeAbrechnungseinheit { get; set; }
 
         public decimal StueckpreisNetto { get; set; }
@@ -27,8 +25,6 @@ namespace DoePaAdminDataModel.Kostenrechnung
                 return StueckpreisNetto * Stueckzahl;
             }
         }
-
-        public Waehrung NettobetragWaehrung { get; set; }
 
         public decimal Steuersatz { get; set; }
 
