@@ -12,7 +12,9 @@ namespace DoePaAdminDataModel.Stammdaten
         public string Anschrift { get; set; }
 
         public Adresse ZugehoerigeAdresse { get; set; }
-                
+
+        public Dictionary<string, string> Additions { get; set; }
+        
         public override string ToString()
         {
             return Anschrift;
