@@ -1,6 +1,6 @@
 ﻿using DoePaAdmin.ViewModel.Services;
 using DoePaAdminDataModel.Stammdaten;
-using Microsoft.Toolkit.Mvvm.Input;
+using CommunityToolkit.Mvvm.Input;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
